docker: Add Fedora support to install_basics.sh

+++ tools/docker/scripts/install_basics.sh
@@ -2,20 +2,42 @@
 
 # author: Ole Schuett
 
-# install Ubuntu packages
-apt-get update -qq
-apt-get install -qq --no-install-recommends \
-    ca-certificates                         \
-    locales                                 \
-    git                                     \
-    less                                    \
-    nano                                    \
-    rsync                                   \
-    wget
+if grep -q "Ubuntu" /etc/os-release ; then
+    echo -n "Installing Ubuntu packages... "
+    apt-get update -qq
+    apt-get install -qq --no-install-recommends \
+        ca-certificates                         \
+        locales                                 \
+        git                                     \
+        less                                    \
+        nano                                    \
+        rsync                                   \
+        wget
+    rm -rf /var/lib/apt/lists/*
+    locale-gen "en_US.UTF-8"
+    echo "done."
 
-rm -rf /var/lib/apt/lists/*
+elif grep -q "Fedora" /etc/os-release ; then
+    echo -n "Installing Fedora packages... "
+    dnf -qy install                     \
+        ca-certificates                 \
+        git                             \
+        less                            \
+        nano                            \
+        rsync                           \
+        wget
+    dnf -q clean all
+    echo "done."
 
-locale-gen "en_US.UTF-8"
+else
+    echo "Unknown Linux distribution."
+    exit 1
+
+fi
 
 # clone cp2k repository
+echo -n "Cloning cp2k repository... "
 git clone --quiet --recursive --single-branch -b master https://github.com/cp2k/cp2k.git /workspace/cp2k
+echo "done."
+
+#EOF
